[shell] Add basic autocompletion support
This change adds autocompletion support for simple statements. It
leverages this lexer from `raccolta-syntax` to validate the input.

In the future, it might be nice to parse statements so it gets it's
suggestions from the errors that arise from there. Mayor rewriting is
needed though to support this, including a list of expected tokens
instead of the token that was found (or EOF).