[syntax] Add `NonReservedWord` keyword type
This change is the first in a series which migrates the existing Keyword
enum into one with non-reserved words and reserved-words.

As of now, `Keyword` incorrectly contains a list of both
non-reserved words and reserved-words. This is a problem, because
non-reserved words contains a list of words that sometimes have semantic
meaning, but are otherwise commonly used as table and column names. (For
example, the non-reserved word `NUMBER`).