[engine] Refactor `SELECT` processing into a streamlined path
This change refactors the processing of `SELECT` statements in a
streamlined path, meaning the `SELECT *` and `SELECT {<column name>,...}`
both return an iterator without ordering.

This makes it so that we can filter, group, and then order without much
code duplication. This was previously already possible, but is now much
better.