[syntax] Mark `StatementParseError`s where delimiters should've been matched
To aid the user in finding where a specific error went wrong, the
position of the opening delimiter is now provided in some error
variants.

The shell can use this information to visually display in the terminal
where the error occurred, alongside the opening delimiter that was
expected to be matched.

CREATE TABLE numbers (number INT
                     ^ the position of this delimiter and the token kind
		       are now stored as well!