[shell] Display delimiter to match in errors
Certain errors signify a condition wherein a closing delimiter is
expected, but is absent. This change adds visual indications of the
position in the statement where such an error ocurred and where the
opening delimiter was present.

See the following example:

===== Input =====
> CREATE TABLE numbers (number INT

===== Output =====
Error(TableElementsUnexpectedEndOfFile): unexpected end-of-file in the table column definitions

  CREATE TABLE numbers (number INT
                       ^         ^
                       |         | error occurred here
                       | match this opening parenthesis here